llamaocr/requirements.txt
groq==0.4.1
pillow==10.0.0
python-dotenv==1.0.0
base64
json
os